third_party/WebKit/LayoutTests/inspector/sources/debugger/source-frame-expected.txt

CONSOLE ERROR: line 12: test error message
Tests that it's possible to set breakpoint in source frame, and that source frame displays breakpoints and console errors.


Running: testSetBreakpoint
Script source was shown.
Breakpoint added to source frame at line 14

Running: testConsoleMessage
Script source was shown.
Message added to source frame: test error message
Messages cleared in source frame.

Running: testShowResource
